[feat]: added __init__ and segregated DB and embedding service
