+++ backend/app/services/vector_db/service.py
@@ -1,8 +1,6 @@
 import logging
 import os
 from typing import List, Dict, Any, Optional
-from uuid import UUID
-from pydantic import BaseModel, Field
 from dotenv import load_dotenv
 from supabase import create_client, Client
 
@@ -12,13 +10,6 @@ supabase_url = os.getenv("SUPABASE_URL")
 supabase_key = os.getenv("SUPABASE_SERVICE_ROLE_KEY")
 
 logger = logging.getLogger(__name__)
-
-class EmbeddingItem(BaseModel):
-    id: str
-    collection: str
-    content: str
-    metadata: Dict[str, Any] = Field(default_factory=dict)
-    embedding: Optional[List[float]] = None
 
 
 class VectorDBService:
